Fix #5 issue by comparing all changes date
Old changes date and changes date in the zip file are compared for
inclusion and not for most recent date. So far future change
and near future change are detected if they are not present
in the current version.txt file. Idem for version tag.
More logging.Debug are added, PEP8 coding style violation
are corrected, and spelling errors corrected.